move functions out of gl
