+++ Engine/Graphics/Direct3D/Render.cpp
@@ -16,7 +16,7 @@
 #include "Engine/Graphics/ImageLoader.h"
 #include "Engine/Graphics/Level/Decoration.h"
 #include "Engine/Graphics/LightmapBuilder.h"
-#include "Engine/Graphics/Lights.h"
+#include "Engine/Graphics/LightsStack.h"
 #include "Engine/Graphics/Outdoor.h"
 #include "Engine/Graphics/PCX.h"
 #include "Engine/Graphics/PaletteManager.h"
